requirements.txt
# Core AI
ollama>=0.1.7
numpy>=1.24.0
requests>=2.31.0

# Web & API
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
websockets>=11.0.0

# Audio Processing
openai-whisper>=20231117
piper-tts>=0.0.2
pyaudio>=0.2.11
soundfile>=0.12.0
speechrecognition>=3.10.0

# Memory & Vectors
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Discord
discord.py>=2.3.0

# Database (REPLACED MySQL-python with modern alternative)
mysql-connector-python>=8.0.0

# Utilities
python-dotenv>=1.0.0
psutil>=5.9.0
colorama>=0.4.6
asyncio>=3.4.3

# Advanced Memory & Learning
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Advanced Features
scikit-learn>=1.3.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Mobile & Streaming
flask-socketio>=5.3.0
eventlet>=0.33.0
opencv-python>=4.8.0
pillow>=10.0.0

# Screen capture for VTube Studio streaming
pyautogui>=0.9.54
pygetwindow>=0.0.9

# Enhanced WebUI
plotly>=5.15.0
dash>=2.14.0
visdcc>=0.0.43
animate.css>=4.0.0
chart.js>=3.9.0